apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic[email]==2.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
Pillow==10.3.0
numpy==1.26.4
tensorflow==2.16.1
tf-keras==2.16.0
opencv-python-headless==4.10.0.82
matplotlib==3.9.0
